requirements.txt
Django~=5.1.6
pip~=24.3.1
djangorestframework~=3.16.0
